requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-countries==7.3.2
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
jsondiff==2.0.0
lxml==4.9.1
Markdown==3.4.1
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.3
pycountry==22.3.5
pycparser==2.21
PyJWT==2.4.0
pymemcache==3.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.2.1
pyuca==1.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
wincertstore==0.2
zipp==3.8.1
